+++ examples/divider.rs
@@ -1,4 +1,4 @@
-use cliux::components::Divider;
+use cliux::Divider;
 
 fn main() {
     Divider::new(20).print();