+++ src/components/mod.rs
@@ -1,9 +1,13 @@
 pub mod boxed;
 pub mod divider;
-pub mod section;
 pub mod label;
+pub mod list;
+pub mod section;
+pub mod tag;
 
 pub use boxed::Boxed;
 pub use divider::Divider;
-pub use section::Section;
 pub use label::Label;
+pub use list::List;
+pub use section::Section;
+pub use tag::Tag;